+++ src/data/repositories/DataSetD2Repository.ts
@@ -4,9 +4,9 @@ import { DataSet, DataSetAttrs } from "$/domain/entities/DataSet";
 import { DataSetRepository } from "$/domain/repositories/DataSetRepository";
 import { D2Api } from "$/types/d2-api";
 import { apiToFuture, FutureData } from "$/data/api-futures";
-import { filterValidInstances } from "$/utils/instance-utils";
 import { Future } from "$/domain/entities/generic/Future";
 import _c from "$/domain/entities/generic/Collection";
+import { filterValidInstances } from "$/domain/entities/generic/Struct";
 
 export class DataSetD2Repository implements DataSetRepository {
     constructor(private api: D2Api) {}
