+++ src/domain/entities/__tests__/BasicDataSet.spec.ts
@@ -10,11 +10,13 @@ describe("BasicDataSet", () => {
     it("valid attributes should return an instance of BasicDataSet", () => {
         expect(validBasic).toBeInstanceOf(BasicDataSet);
     });
+
     it("should throw an error when the BasicDataSet has a formType equal to CUSTOM", () => {
         const basicDataSet = () => new BasicDataSet(basicAttrsWithCustomFormType());
 
         expect(basicDataSet).toThrowError("DataSet BwyMfDBLih9 has CUSTOM form type");
     });
+
     it("should throw an error when the BasicDataSet has an attribute with name hideInTallySheet and value true", () => {
         const basicDataSet = () => new BasicDataSet(basicAttrsWithHideInTallySheetsAttribute());
