+++ src/domain/entities/__tests__/User.spec.ts
@@ -7,11 +7,13 @@ describe("User", () => {
 
         expect(user.isAdmin()).toBe(true);
     });
+
     it("should no be admin if hasn't a role with authority ALL", () => {
         const user = createNonAdminUser();
 
         expect(user.isAdmin()).toBe(false);
     });
+
     it("should return belong to user group equal to false when the id exist", () => {
         const userGroupId = "BwyMfDBLih9";
 
@@ -19,6 +21,7 @@ describe("User", () => {
 
         expect(user.belongToUserGroup(userGroupId)).toBe(true);
     });
+
     it("should return belong to user group equal to false when the id does not exist", () => {
         const existedUserGroupId = "BwyMfDBLih9";
         const nonExistedUserGroupId = "f31IM13BgwJ";
@@ -27,6 +30,7 @@ describe("User", () => {
 
         expect(user.belongToUserGroup(nonExistedUserGroupId)).toBe(false);
     });
+
     it("should return belong to user group equal to false if user groups is empty", () => {
         const nonExistedUserGroupId = "f31IM13BgwJ";
