+++ src/domain/entities/generic/Struct.ts
@@ -43,3 +43,32 @@ export function Struct<Attrs>() {
 const GenericStruct = Struct<unknown>();
 
 export type GenericStructInstance = InstanceType<typeof GenericStruct>;
+
+interface InstancesWithErrors<InstanceType> {
+    instances: InstanceType[];
+    errors: Error[];
+}
+
+export function filterValidInstances<Attrs, InstanceType>(
+    Instance: { new (attrs: Attrs): InstanceType },
+    attrs: Attrs[]
+): InstancesWithErrors<InstanceType> {
+    return attrs.reduce<InstancesWithErrors<InstanceType>>(
+        (acc, attr) => {
+            try {
+                // Using .push() to avoid creating new array instances for performance
+                acc.instances.push(new Instance(attr));
+            } catch (e) {
+                if (e instanceof Error) {
+                    console.debug(e.message);
+                    acc.errors.push(e);
+                } else {
+                    console.error("Unknown error: ", e);
+                }
+            }
+
+            return acc;
+        },
+        { instances: [], errors: [] }
+    );
+}
